enabled build from source dir/tarball via  to --openjdk-source-location/-l (#3737)
* enabled build from source dir/tarball via  to --openjdk-source

* not excluding build from tarballs extraction, rmeoving instead

* removed debug

* Fixing issues found during review

redundant if [ "${BUILD_CONFIG[OPENJDK_FOREST_DIR]}" == "true" ] from getFirstTagFromOpenJDKGitRepo
renamed setLocalDir to setOpenjdkSourceDir
wrong parent of  removed build dir after unpack
bounch of typos:
  does not exists
  Copying OpenJDK source from
  Extracting OpenJDK source tarbal
  contains, does not contain

* Quoting few variables to prevent globbing and word splitting

thanx shell lint

* Replaced `pwd` by $(pwd)

* Update makejdk-any-platform.1


* Update makejdk-any-platform.1


* Update makejdk-any-platform.1


* Update sbin/prepareWorkspace.sh


* Update sbin/prepareWorkspace.sh


* Update sbin/prepareWorkspace.sh


* Update sbin/prepareWorkspace.sh


* Update makejdk-any-platform.1


* removed useless --strip-components 0 to help a it non posix systems

* be specific on on kind of git ops hidden below magic

* Using verbose sleep instead of jsut sleep

* Returned globing for $files, disable=SC2086 for that line

* sed "s/OPENJDK_FOREST_DIR/OPENJDK_FOREST_SOURCE_ARCHIVE/g" -i {prepareWorkspace.sh,config_init.sh,build.sh}

* Update makejdk-any-platform.1


* Update sbin/build.sh


* Update sbin/build.sh


* Update sbin/build.sh


* Update sbin/common/config_init.sh


* Update sbin/common/config_init.sh


* Update sbin/prepareWorkspace.sh


* Update sbin/common/config_init.sh


* Update sbin/prepareWorkspace.sh


* Fixed last ;shift; to ; shift;

* Update sbin/prepareWorkspace.sh


* Update sbin/prepareWorkspace.sh


* Extracting OpenJDK source tarball ... to build the binary

resolving: https://github.com/adoptium/temurin-build/pull/3737/files/38e318f2dc0234a2a4789510280be0e34a7b3d59#r1548826320

* Update sbin/prepareWorkspace.sh


* You have specified ... does not exist

resolved: https://github.com/adoptium/temurin-build/pull/3737/files/38e318f2dc0234a2a4789510280be0e34a7b3d59#r1548741145

* using OPENJDK_LOCAL_SOURCE_ARCHIVE instead of OPENJDK_FOREST_SOURCE_ARCHIVE

Tus resulting for both usages:
  -OPENJDK_FOREST_SOURCE_ARCHIVE
  -OPENJDK_FOREST_SOURCE_ARCHIVE_ABSPATH
  +OPENJDK_LOCAL_SOURCE_ARCHIVE
  +OPENJDK_LOCAL_SOURCE_ARCHIVE_ABSPATH

* Update sbin/common/config_init.sh


* Update sbin/build.sh


* Update makejdk-any-platform.1


* Update makejdk-any-platform.1


* Update makejdk-any-platform.1


* Made tar --exclude check nonfatal

* SC2126 Considered using 'grep -c' instead of 'grep|wc -l'.

* Workaroudned usage of --strip-components 1 by custom function

usabel also for zips if needed

Added fogooten quotes

* renamed from  "--openjdk-source" | "-o" to  "--openjdk-source-location" | "-l"

* Using $() isntead of backticks

* Moved -l/openjdk-source-location  to proper palce in man after rename from -o

---------

+++ sbin/prepareWorkspace.sh
@@ -22,7 +22,6 @@
 ################################################################################
 
 set -eu
-
 SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
 # shellcheck source=sbin/common/constants.sh
@@ -38,6 +37,64 @@ ALSA_LIB_CHECKSUM=${ALSA_LIB_CHECKSUM:-5f2cd274b272cae0d0d111e8a9e363f0878332915
 ALSA_LIB_GPGKEYID=${ALSA_LIB_GPGKEYID:-A6E59C91}
 FREETYPE_FONT_SHARED_OBJECT_FILENAME="libfreetype.so*"
 
+
+copyFromDir() {
+  echo "Copying OpenJDK source from  ${BUILD_CONFIG[OPENJDK_LOCAL_SOURCE_ARCHIVE_ABSPATH]} to $(pwd)/${BUILD_CONFIG[OPENJDK_SOURCE_DIR]} to be built"
+  # We really do not want to use .git for dirs, as we expect user have them set up, ignoring them
+  local files=$(find "${BUILD_CONFIG[OPENJDK_LOCAL_SOURCE_ARCHIVE_ABSPATH]}" -maxdepth 1 -mindepth 1 | grep -v -e "/workspace$" -e "/build$" -e "/.git" -e -"/build/")
+  # SC2086 (info): Double quote to prevent globbing and word splitting.
+  # globbing is intentional here
+  # shellcheck disable=SC2086
+  cp -rf $files "./${BUILD_CONFIG[OPENJDK_SOURCE_DIR]}/"
+}
+
+# this is workarounding --strip-components 1 missing on gnu tar
+# it requires  absolute tar-filepath as it changes dir and is hardcoded to one
+# similar approach can be used also for zip in future
+untarGnuAbsPathWithStripComponents1() {
+  local tmp=$(mktemp -d)
+  pushd "$tmp" > /dev/null
+    tar "$@"
+  popd  > /dev/null
+  mv "$tmp"/*/* .
+  mv "$tmp"/*/.* . || echo "no hidden files in tarball"
+  rmdir "$tmp"/*
+  rmdir "$tmp"
+}
+
+unpackFromArchive() {
+  echo "Extracting OpenJDK source tarball ${BUILD_CONFIG[OPENJDK_LOCAL_SOURCE_ARCHIVE_ABSPATH]} to $(pwd)/${BUILD_CONFIG[OPENJDK_SOURCE_DIR]} to build the binary"
+  # If the tarball contains .git files, they should be ignored later
+  # todo, support also zips?
+  pushd "./${BUILD_CONFIG[OPENJDK_SOURCE_DIR]}"
+    local topLevelItems=$(tar --exclude='*/*' -tf  "${BUILD_CONFIG[OPENJDK_LOCAL_SOURCE_ARCHIVE_ABSPATH]}"  | grep "/$" -c) || local topLevelItems=1
+    if [ "$topLevelItems" -eq "1" ] ; then
+      echo "Source tarball contains exactly one directory"
+      untarGnuAbsPathWithStripComponents1 -xf "${BUILD_CONFIG[OPENJDK_LOCAL_SOURCE_ARCHIVE_ABSPATH]}"
+    else
+      echo "Source tarball does not contain a top level directory"
+      tar -xf "${BUILD_CONFIG[OPENJDK_LOCAL_SOURCE_ARCHIVE_ABSPATH]}"
+    fi
+    rm -rf "build"
+  popd
+}
+
+copyFromDirOrUnpackFromArchive() {
+  echo "Cleaning the copy of OpenJDK source repository from $(pwd)/${BUILD_CONFIG[OPENJDK_SOURCE_DIR]} and replacing with a fresh copy in 10 seconds..."
+  verboseSleep	 10
+  rm -rf "./${BUILD_CONFIG[OPENJDK_SOURCE_DIR]}"
+  mkdir  "./${BUILD_CONFIG[OPENJDK_SOURCE_DIR]}"
+  # Note that we are not persisting the build directory
+  if [ -d "${BUILD_CONFIG[OPENJDK_LOCAL_SOURCE_ARCHIVE_ABSPATH]}" ] ; then
+    copyFromDir
+  elif [ -f "${BUILD_CONFIG[OPENJDK_LOCAL_SOURCE_ARCHIVE_ABSPATH]}" ] ; then
+    unpackFromArchive
+  else
+    echo "${BUILD_CONFIG[OPENJDK_LOCAL_SOURCE_ARCHIVE_ABSPATH]} is not a directory or a file "
+    exit 1
+  fi
+}
+
 # Create a new clone or update the existing clone of the OpenJDK source repo
 # TODO refactor this for Single Responsibility Principle (SRP)
 checkoutAndCloneOpenJDKGitRepo() {
@@ -45,7 +102,7 @@ checkoutAndCloneOpenJDKGitRepo() {
   cd "${BUILD_CONFIG[WORKSPACE_DIR]}/${BUILD_CONFIG[WORKING_DIR]}"
 
   # Check that we have a git repo, we assume that it is a repo that contains openjdk source
-  if [ -d "${BUILD_CONFIG[OPENJDK_SOURCE_DIR]}/.git" ]; then
+  if [ -d "${BUILD_CONFIG[OPENJDK_SOURCE_DIR]}/.git" ] && [ "${BUILD_CONFIG[OPENJDK_LOCAL_SOURCE_ARCHIVE]}" == "false" ]; then
     set +e
     git --git-dir "${BUILD_CONFIG[OPENJDK_SOURCE_DIR]}/.git" remote -v
     echo "${BUILD_CONFIG[OPENJDK_CORE_VERSION]}"
@@ -79,6 +136,8 @@ checkoutAndCloneOpenJDKGitRepo() {
       echo "If this is inside a docker you can purge the existing source by passing --clean-docker-build"
       exit 1
     fi
+  elif [ "${BUILD_CONFIG[OPENJDK_LOCAL_SOURCE_ARCHIVE]}" == "true" ]; then
+    copyFromDirOrUnpackFromArchive
   elif [ ! -d "${BUILD_CONFIG[OPENJDK_SOURCE_DIR]}/.git" ]; then
     echo "Could not find a valid openjdk git repository at $(pwd)/${BUILD_CONFIG[OPENJDK_SOURCE_DIR]} so re-cloning the source to openjdk"
     rm -rf "${BUILD_CONFIG[WORKSPACE_DIR]:?}/${BUILD_CONFIG[WORKING_DIR]}/${BUILD_CONFIG[OPENJDK_SOURCE_DIR]}"
@@ -107,8 +166,9 @@ checkoutAndCloneOpenJDKGitRepo() {
     fi
   fi
 
-  git clean -ffdx
-
+  if [ "${BUILD_CONFIG[OPENJDK_LOCAL_SOURCE_ARCHIVE]}" == "false" ]; then
+    git clean -ffdx
+  fi
   updateOpenj9Sources
 
   createSourceTagFile
@@ -119,6 +179,17 @@ checkoutAndCloneOpenJDKGitRepo() {
 # Checkout the required code to build from the given cached git repo
 # Set checkoutRc to result so we can retry
 checkoutRequiredCodeToBuild() {
+
+  if [ "${BUILD_CONFIG[OPENJDK_LOCAL_SOURCE_ARCHIVE]}" == "true" ]; then
+    echo "Skipping checkoutRequiredCodeToBuild - local directory under processing:"
+    echo "  workspace = ${BUILD_CONFIG[WORKSPACE_DIR]}/${BUILD_CONFIG[WORKING_DIR]}/${BUILD_CONFIG[OPENJDK_SOURCE_DIR]}"
+    echo "  BUILD_VARIANT = ${BUILD_CONFIG[BUILD_VARIANT]}"
+    echo "  TAG = ${BUILD_CONFIG[TAG]} - Used only in name, if at all"
+    echo "  BRANCH = ${BUILD_CONFIG[BRANCH]} - UNUSED!"
+    checkoutRc=0
+    return
+  fi
+
   checkoutRc=1
 
   cd "${BUILD_CONFIG[WORKSPACE_DIR]}/${BUILD_CONFIG[WORKING_DIR]}/${BUILD_CONFIG[OPENJDK_SOURCE_DIR]}"
